refactor: rename schema to typeDefs
